refactor(library): split mod.rs into focused submodules
Decompose the 3725-line library/mod.rs into 7 domain-specific
submodules to improve incremental compile times and readability:

- lightbox.rs: full-screen image viewer, zoom, pan, EXIF details (919 LOC)
- download.rs: download flow, video handoff, transfer tracking (289 LOC)
- actions.rs: selection-mode wiring and bulk batch operations (198 LOC)
- context_menu.rs: right-click popover and action handlers (243 LOC)
- album_link.rs: album-folder linking, sync dialog (312 LOC)
- controls.rs: header bar, search, sort, sidebar/grid handlers (386 LOC)
- filters.rs: advanced metadata-filters dialog (264 LOC)

mod.rs retains the LibraryWindowUi struct, build_library_window,
data-loading dispatch, and transfer UI (1232 LOC).

All new modules use pub(super) visibility -- zero external import
changes required. All tests pass, clippy clean.
